feat(ui): wip mutualisation filtres
